fixes issue with critical facilites scaling
critical facilites scale is 1 and/or 5, but zonal stats could be 1-5.  The legend is drawn from the mapconfig.js file, and uses the keys chartLegendValues, chartCSSColor. The chart creation and chart color needed 1-5 values.  but since the key chartMaxValue was 5, and it used and index to determine th e chartCSSColor and unkown index color was thrown.  this should resolve it. made intermediate colors 2,3,4,5 the same as 5 so the legend "looks" correct.